requirements.txt
Django==4.2.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ipython==8.28.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
stripe
eventlet
redis
django-celery-beat
django-cors-headers
drf_yasg